Moved tokenizers to a dedicated folder
