+++ src/python/CMakeLists.txt
@@ -18,10 +18,10 @@ if(NOT pybind11_POPULATED)
     add_subdirectory(${pybind11_SOURCE_DIR} ${pybind11_BINARY_DIR})
 endif()
 
-pybind11_add_module(py_generate_pipeline py_vlm_pipeline.cpp py_generate_pipeline.cpp py_whisper_pipeline.cpp py_utils.cpp)
-target_link_libraries(py_generate_pipeline PRIVATE openvino::genai)
-target_include_directories(py_generate_pipeline PRIVATE "${OpenVINOGenAI_SOURCE_DIR}/src/cpp/src") # for tokenizers_path.hpp
-set_target_properties(py_generate_pipeline PROPERTIES
+pybind11_add_module(py_openvino_genai py_vlm_pipeline.cpp py_llm_pipeline.cpp py_whisper_pipeline.cpp py_utils.cpp)
+target_link_libraries(py_openvino_genai PRIVATE openvino::genai)
+target_include_directories(py_openvino_genai PRIVATE "${OpenVINOGenAI_SOURCE_DIR}/src/cpp/src") # for tokenizers_path.hpp
+set_target_properties(py_openvino_genai PROPERTIES
     ARCHIVE_OUTPUT_DIRECTORY "$<1:${CMAKE_BINARY_DIR}/openvino_genai/>"
     LIBRARY_OUTPUT_DIRECTORY "$<1:${CMAKE_BINARY_DIR}/openvino_genai/>"
 )
@@ -51,14 +51,14 @@ elseif(APPLE)
 endif()
 
 if(rpaths)
-    set_target_properties(py_generate_pipeline PROPERTIES INSTALL_RPATH "${rpaths}")
+    set_target_properties(py_openvino_genai PROPERTIES INSTALL_RPATH "${rpaths}")
 endif()
 
 install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/openvino_genai/__init__.py"
               "${CMAKE_BINARY_DIR}/openvino_genai/__version__.py"
         DESTINATION python/openvino_genai
         COMPONENT pygenai_${Python3_VERSION_MAJOR}_${Python3_VERSION_MINOR})
-install(TARGETS py_generate_pipeline
+install(TARGETS py_openvino_genai
         LIBRARY DESTINATION python/openvino_genai
         COMPONENT pygenai_${Python3_VERSION_MAJOR}_${Python3_VERSION_MINOR})
 
@@ -76,6 +76,6 @@ install(FILES "${OpenVINOGenAI_SOURCE_DIR}/LICENSE"
 
 # wheel_genai component is used for wheel generation in pyproject.toml.
 # Exclude wheel_genai from normal packaging because there's pygenai_X_Y component for that.
-install(TARGETS openvino_genai py_generate_pipeline
+install(TARGETS openvino_genai py_openvino_genai
         LIBRARY DESTINATION openvino_genai COMPONENT wheel_genai EXCLUDE_FROM_ALL
         RUNTIME DESTINATION openvino_genai COMPONENT wheel_genai EXCLUDE_FROM_ALL)
