+++ src/python/py_whisper_pipeline.cpp
@@ -5,7 +5,7 @@
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
 
-#include "tokenizers_path.hpp"
+#include "tokenizer/tokenizer_path.hpp"
 #include "openvino/genai/whisper_generation_config.hpp"
 #include "openvino/genai/whisper_pipeline.hpp"
 #include "py_utils.hpp"
